fix:#03 solved gsap is other file
