+++ src/001_dev_style/scripts/bootstrap.js
@@ -6,13 +6,14 @@ import { loader } from "./component/loader";
 import { theme } from "./component/theme";
 import { gui } from "./helper/gui";
 import menu from "./component/menu";
-import { cling } from "./component/cling";
+import { cling } from "./component/clingHeader";
 import { homeNews } from "./component/home-news";
 import "./component/scroll-animation";
 import { iNode } from "./helper";
 
 import { elementPos } from "./component/elementpos";
 import { elementPosHome } from "./component/elementposHome";
+
 
 window.debug = debugmode(0) ? 1 : 0;
