+++ src/001_dev_style/scripts/myclasses/header.js
@@ -16,7 +16,12 @@ class HeaderHandler {
     this.headerLogo = iNode.qs(".Header__Icon");
     this.headerMainNav = iNode.qs(".Header__MainNav");
     this.secondNav = iNode.qs(".Header__secondaryNav");
+    this.fv = iNode.qs(".fv");
+    this.homeNews = iNode.qs(".home-news");
+    this.footer = iNode.qs(".Footer");
 
+    this.headerWrap = iNode.qs(".Header__Wrap");
+    this.sectionTemplate = iNode.qs(".section--home-panels");
     this.logoGray = iNode.qs(".Logo__gray");
     this.logoWhite = iNode.qs(".Logo__white");
     this.btnBars = iNode.qsa(".btn-menu_bar");
@@ -27,6 +32,8 @@ class HeaderHandler {
     this.HeaderFlex = iNode.qs(".Header__FlexItem--logo");
     this.btn = iNode.qs(".btn-menu");
     this.HeaderIcon = iNode.qs(".Header__Icon");
+    this.clingToStyleOnEnter();
+    this.clingToStyleOnLeaveBack();
   }
 
   isWideToggler(isWideScreen) {
@@ -69,29 +76,46 @@ class HeaderHandler {
   }
 
   setElHeight(value) {
-    // console.log("value", value);
-    // const header = iNode.qs("#header");
-    // const headerClass = iNode.qs(".Header");
-
-    // headerClass.style.height = `${value}px`;
-    // headerClass.style.maxHeight = `${value}px`;
-
     this.header.style.height = `${value}px`;
     this.header.style.maxHeight = `${value}px`;
-    console.log("this.header", this.header);
+    // console.log("this.header", this.header);
 
-    // this.header = header;
-    // console.log("this.header.height", this.header.style.height);
+    const pinSpacer = iNode.qs(".pin-spacer");
+    // console.log("pinSpacer", pinSpacer);
+    if (!pinSpacer) return this.header;
 
-    const test = iNode.qs(".pin-spacer");
-    console.log("test", test);
-    if (!test) return this.header;
-
-    if (test) {
-      test.style.height = `${value}px`;
-      test.style.maxHeight = `${value}px`;
+    if (pinSpacer) {
+      pinSpacer.style.height = `${value}px`;
+      pinSpacer.style.maxHeight = `${value}px`;
     }
     // return this.header;
+  }
+
+  clingToStyleOnEnter() {
+    iNode.toggleClass(this.headerWrap, "Header--white", true);
+    // iNode.toggleClass(header, "Header--white", true);
+
+    iNode.setStyles(this.logoGray, { opacity: 1 });
+    iNode.setStyles(this.logoWhite, { opacity: 0 });
+    this.btnBars.forEach((btnBar) => {
+      iNode.setStyles(btnBar, { backgroundColor: "var(--color-gray)" });
+    });
+    iNode.setStyles(this.cart, { color: "var(--color-gray)" });
+    iNode.setStyles(this.headerMainNav, { color: "var(--color-gray)" });
+    iNode.setStyles(this.secondNav, { color: "var(--color-gray)" });
+  }
+
+  clingToStyleOnLeaveBack() {
+    iNode.toggleClass(this.headerWrap, "Header--white", false);
+    // iNode.toggleClass(header, "Header--white", false);
+    iNode.setStyles(this.logoGray, { opacity: 0 });
+    iNode.setStyles(this.logoWhite, { opacity: 1 });
+    this.btnBars.forEach((btnBar) => {
+      iNode.setStyles(btnBar, { backgroundColor: "var(--color-border)" });
+    });
+    iNode.setStyles(this.cart, { color: "var(--color-border)" });
+    iNode.setStyles(this.headerMainNav, { color: "var(--color-border)" });
+    iNode.setStyles(this.secondNav, { color: "var(--color-border)" });
   }
 }
