+++ src/001_dev_style/scripts/page/home.js
@@ -57,6 +57,7 @@ export default async function init({
     },
   });
 
+
   // elementPos.init();
   // elementPos.resizeHeaderPos();
   // elementPos.resizingFooterPos();
